Provide default methods for rescale and rescale_mid
Closes #102.

This commit provides a default method to rescale and rescale_mid. The default method coerces the data to numeric, as this is the right thing to do in most cases. It also gives a warning, so the scales user is aware of the coercion (it is not always the right thing to do).